# api/openapi.yaml
openapi: 3.1.0
info:
  title: Speccy API
  version: 0.1.0
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
paths:
  /meta:
    get:
      operationId: getMeta
      summary: Get the server version and mode.
      responses:
        "200":
          description: The server version and mode.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meta"
        default:
          $ref: "#/components/responses/Problem"
...
  /share/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getShare
      summary: Look up a share link (REQ-085). A revoked or expired link is not found.
      responses:
        "200":
          description: The shared bundle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareInfo"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: joinShare
      summary: Enter a share link as a guest with a display name (REQ-086). Sets the guest cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [display_name]
              properties:
                display_name:
                  type: string
                  minLength: 1
                  maxLength: 60
      responses:
        "200":
          description: The guest may now read the bundle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareInfo"
        default:
          $ref: "#/components/responses/Problem"
  /bundles/{bundleId}/access:
    parameters:
      - $ref: "#/components/parameters/BundleId"
    get:
      operationId: getBundleAccess
      summary: Who can see the bundle, and its share link state (REQ-084, REQ-085).
      responses:
        "200":
          description: The access of the bundle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleAccess"
        default:
          $ref: "#/components/responses/Problem"
  /handoffs/{handoffId}/report:
    parameters:
      - $ref: "#/components/parameters/HandoffId"
    post:
      operationId: reportBuild
      summary: Report what the build learned about the doc (REQ-137). A blocked report opens a blocking thread.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildReport"
      responses:
        "200":
          description: The thread the report opened.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadDetail"
        default:
          $ref: "#/components/responses/Problem"
  /docs/{docId}/handoff:
    parameters:
      - $ref: "#/components/parameters/DocId"
    get:
      operationId: listHandoffs
      summary: The handoffs of a bundle, newest first (REQ-136).
      responses:
        "200":
          description: The handoffs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandoffList"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: takeHandoff
      summary: Take the build packet of a Build Ready bundle, and record the handoff (REQ-136).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                acknowledged:
                  type: boolean
                  description: Take the packet although the verdict is not Build Ready, or is stale. The handoff records the verdict it was taken at.
                label:
                  type: string
                  description: What the builder calls this work, such as a repo, a branch, or a ticket.
      responses:
        "200":
          description: The build packet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildPacket"
        default:
          $ref: "#/components/responses/Problem"
  /docs/{docId}/verification-waivers:
    parameters:
      - $ref: "#/components/parameters/DocId"
    post:
      operationId: requestVerificationWaiver
      summary: Ask to excuse one trace ID in one code repo. It never goes in the doc's sidecar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trace_id, repo, reason]
              properties:
                trace_id:
                  type: string
                repo:
                  type: string
                  description: The code repo, or the folder, this excuse applies to.
                reason:
                  type: string
                  minLength: 20
      responses:
        "200":
          description: The waiver.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Waiver"
        default:
          $ref: "#/components/responses/Problem"
  /bundles/{bundleId}/delete-plan:
    parameters:
      - $ref: "#/components/parameters/BundleId"
    get:
      operationId: deleteBundlePlan
      summary: What the Delete control offers for this bundle, by the kind of source that makes it.
      responses:
        "200":
          description: The plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletePlan"
        default:
          $ref: "#/components/responses/Problem"
  /docs/{docId}/verifications:
    parameters:
      - $ref: "#/components/parameters/DocId"
    get:
      operationId: listVerifications
      summary: The verification runs of a bundle, newest first.
      responses:
        "200":
          description: The verification runs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationList"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: runVerification
      summary: Queue a verification of one build of this bundle against a code repo at one commit, or a folder.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationRequest"
      responses:
        "202":
          description: The queued run. Follow its progress at /verifications/{runId}/events, or read it until its status is done or failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verification"
        default:
          $ref: "#/components/responses/Problem"
  /docs/{docId}/verifications/resolve:
    parameters:
      - $ref: "#/components/parameters/DocId"
    post:
      operationId: resolveVerificationTarget
      summary: Say which repo and commit, or which folder, a pasted target names, before a run starts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationTargetInput"
      responses:
        "200":
          description: The build the target names.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolvedBuild"
        default:
          $ref: "#/components/responses/Problem"
  /docs/{docId}/verifications/defaults:
    parameters:
      - $ref: "#/components/parameters/DocId"
    get:
      operationId: verificationDefaults
      summary: The targets that prefill the verify field. The implemented-by links first, else the repo of the last run.
      responses:
        "200":
          description: The targets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationDefaults"
        default:
          $ref: "#/components/responses/Problem"
  /verifications/{runId}/events:
    parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: verificationEvents
      summary: Follow a verification run's progress by trace ID, as server-sent events.
      responses:
        "200":
          description: A stream of events. Each data line is a RunEvent. The stream ends when the run ends.
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Problem"
  /verifications/{runId}:
    parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getVerification
      summary: One verification run with the outcome of each trace ID.
      responses:
        "200":
          description: The verification run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verification"
        default:
          $ref: "#/components/responses/Problem"
  /docs/{docId}/adopt:
    parameters:
      - $ref: "#/components/parameters/DocId"
    post:
      operationId: adoptFrontmatter
      summary: Write the type and the size that the review used into the main doc's frontmatter (REQ-135).
      responses:
        "200":
          description: The new version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WriteResult"
        default:
          $ref: "#/components/responses/Problem"
  /docs/{docId}/profile:
    parameters:
      - $ref: "#/components/parameters/DocId"
    put:
      operationId: setBundleProfile
      summary: Change the profile of the bundle's main doc.
      description: >
        A doc Speccy owns, on disk or in the store, gets the type written into its frontmatter as
        a new version. A doc in a repo source gets an adopted type, and the repo takes no commit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile]
              properties:
                profile:
                  type: string
      responses:
        "200":
          description: The bundle, with its new profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecDoc"
        default:
          $ref: "#/components/responses/Problem"
  /bundles/{bundleId}/visibility:
    parameters:
      - $ref: "#/components/parameters/BundleId"
    put:
      operationId: setVisibility
      summary: Set the visibility of the bundle (REQ-084). Leaving link visibility revokes the share link.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [visibility]
              properties:
                visibility:
                  $ref: "#/components/schemas/Visibility"
      responses:
        "200":
          description: The access of the bundle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleAccess"
        default:
          $ref: "#/components/responses/Problem"
  /bundles/{bundleId}/share:
    parameters:
      - $ref: "#/components/parameters/BundleId"
    post:
      operationId: createShareLink
      summary: Make a new share link (REQ-085). It replaces the old one and sets link visibility. The URL appears one time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expires_at:
                  type: string
                  format: date-time
      responses:
        "200":
          description: The new link.
          content:
            application/json:
              schema:
                type: object
                required: [url, access]
                properties:
                  url:
                    type: string
                  access:
                    $ref: "#/components/schemas/BundleAccess"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: revokeShareLink
      summary: Revoke the share link (REQ-085).
      responses:
        "200":
          description: The access of the bundle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleAccess"
        default:
          $ref: "#/components/responses/Problem"
  /admin/invites:
    get:
      operationId: listInvites
      summary: List invite links, newest first (REQ-081).
      responses:
        "200":
          description: The invites. A list never carries a token.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invite"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createInvite
      summary: Make a single-use invite link with a role (REQ-081). The URL appears one time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, member]
      responses:
        "200":
          description: The invite and its link.
          content:
            application/json:
              schema:
                type: object
                required: [url, invite]
                properties:
                  url:
                    type: string
                  invite:
                    $ref: "#/components/schemas/Invite"
        default:
          $ref: "#/components/responses/Problem"
  /admin/invites/{inviteId}/revoke:
    parameters:
      - name: inviteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: revokeInvite
      summary: Revoke an unused invite link.
      responses:
        "204":
          description: Revoked.
        default:
          $ref: "#/components/responses/Problem"
  /admin/reset-links:
    post:
      operationId: createResetLink
      summary: Make a one-time password reset link for a user (REQ-082). The URL appears one time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        "200":
          description: The link.
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url:
                    type: string
        default:
          $ref: "#/components/responses/Problem"
  /admin/settings:
    get:
      operationId: getSettings
      summary: The workspace settings (REQ-009, REQ-081, REQ-105).
      responses:
        "200":
          description: The settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        default:
          $ref: "#/components/responses/Problem"
    put:
      operationId: setSettings
      summary: Change the workspace settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: The settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        default:
          $ref: "#/components/responses/Problem"
  /bundles:
    get:
      operationId: listBundles
      summary: List bundles, and the folders that look like bundles but are not valid.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: One page of bundles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleList"
        default:
          $ref: "#/components/responses/Problem"

    post:
      operationId: createBundle
      summary: Create a bundle from a profile's template (REQ-016).
      description: In local mode, Speccy writes the bundle to a new folder under the folder that Speccy serves.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBundleRequest"
      responses:
        "201":
          description: The new bundle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecDoc"
        default:
          $ref: "#/components/responses/Problem"
  /bundles/import:
    post:
      operationId: importBundle
      summary: Import a bundle from a .md file, a .zip file, or pasted markdown.
      description: >
        Send exactly one of `file` or `text`. In local mode, Speccy writes the bundle to a new
        folder named `name` under the folder that Speccy serves (REQ-008).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImportRequest"
      responses:
        "201":
          description: The new bundles. A folder with two or more spec docs gives one bundle per doc.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleList"
        default:
          $ref: "#/components/responses/Problem"
  /bundles/import/preview:
    post:
      operationId: previewImport
      summary: List the markdown files of an import, with the profile of each and the links Speccy offers.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImportRequest"
      responses:
        "200":
          description: The docs of the import.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreview"
        default:
          $ref: "#/components/responses/Problem"
  /bundles/{bundleId}:
    parameters:
      - $ref: "#/components/parameters/BundleId"
    get:
      operationId: getBundle
      summary: Get one bundle with its spec docs.
      responses:
        "200":
          description: The bundle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteBundle
      summary: Delete a bundle whose text Speccy holds, with everything that hangs off it. Permanent.
      parameters:
        - name: slug
          in: query
          required: true
          description: The bundle's slug, typed to confirm.
          schema: { type: string }
      responses:
        "204":
          description: The bundle is gone.
        default:
          $ref: "#/components/responses/Problem"
  /docs/{docId}:
    parameters:
      - $ref: "#/components/parameters/DocId"
    get:
      operationId: getSpecDoc
      summary: Get one spec doc.
      responses:
        "200":
          description: The spec doc.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecDoc"
        default:
          $ref: "#/components/responses/Problem"
...
  /docs/{docId}/files/rename:
    parameters:
      - $ref: "#/components/parameters/DocId"
    post:
      operationId: renameFile
      summary: Rename or move a file inside the bundle. Creates a version (REQ-005).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameRequest"
      responses:
        "200":
          description: The version that the change created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WriteResult"
        default:
          $ref: "#/components/responses/Problem"
...
  /inbox/seen:
    post:
      operationId: markInboxSeen
      summary: Mark the inbox as read up to now.
      responses:
        "204":
          description: Marked.
        default:
          $ref: "#/components/responses/Problem"
...
components:
  parameters:
...
    BundleId:
      name: bundleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DocId:
      name: docId
      in: path
      required: true
      description: The ID of one spec doc.
      schema:
        type: string
        format: uuid
    HandoffId:
      name: handoffId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: An error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Meta:
      type: object
      additionalProperties: false
      required: [version, mode]
      properties:
        version:
          type: string
          description: The build version of the binary.
        mode:
          type: string
          enum: [local, hosted]
          description: The mode the server runs in.
        sign_in_providers:
          type: array
          description: The OAuth providers of hosted mode (REQ-080), by their auth-all ID. A provider signs in a person who linked it to their account.
          items:
            type: string
            enum: [oidc, github]
...
    Visibility:
      type: string
      enum: [private, internal, link]
    ShareInfo:
      type: object
      required: [bundle_id, title]
      properties:
        bundle_id:
          type: string
          format: uuid
        title:
          type: string
...